requirements.txt
fastapi>=0.109.2
uvicorn>=0.27.1
pydantic>=2.6.1
python-multipart>=0.0.9
replicate>=0.25.1
google-generativeai>=0.8.0
pillow>=10.0.0
